requirements.txt
accelerate==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
diffusers==0.29.2
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
packaging==25.0
pillow==11.0.0
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.3
tokenizers==0.15.2
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.67.1
transformers==4.36.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.4.0
zipp==3.22.0
